+++ langchain/chains/sequential.py
@@ -7,8 +7,8 @@ from pydantic import BaseModel, Extra, root_validator
 from langchain.chains.base import Chain
 
 
-class Pipeline(Chain, BaseModel):
-    """Chain pipeline where the outputs of one step feed directly into next."""
+class SequentialChain(Chain, BaseModel):
+    """Chain where the outputs of one step feed directly into next."""
 
     chains: List[Chain]
     input_variables: List[str]
@@ -76,3 +76,55 @@ class Pipeline(Chain, BaseModel):
                 print(f"\033[1mChain {i}\033[0m:\n{outputs}\n")
             known_values.update(outputs)
         return {k: known_values[k] for k in self.output_variables}
+
+
+class SimpleSequentialPipeline(Chain, BaseModel):
+    """Simple chain where the outputs of one step feed directly into next."""
+
+    chains: List[Chain]
+    input_key: str = "input"  #: :meta private:
+    output_key: str = "output"  #: :meta private:
+
+    class Config:
+        """Configuration for this pydantic object."""
+
+        extra = Extra.forbid
+        arbitrary_types_allowed = True
+
+    @property
+    def input_keys(self) -> List[str]:
+        """Expect input key.
+
+        :meta private:
+        """
+        return [self.input_key]
+
+    @property
+    def output_keys(self) -> List[str]:
+        """Return output key.
+
+        :meta private:
+        """
+        return [self.output_key]
+
+    @root_validator()
+    def validate_chains(cls, values: Dict) -> Dict:
+        """Validate that chains are all single input/output."""
+        for chain in values["chains"]:
+            if len(chain.input_keys) != 1:
+                raise ValueError(
+                    "Chains used in SimplePipeline should all have one input, got "
+                    f"{chain} with {len(chain.input_keys)} inputs."
+                )
+            if len(chain.output_keys) != 1:
+                raise ValueError(
+                    "Chains used in SimplePipeline should all have one output, got "
+                    f"{chain} with {len(chain.output_keys)} outputs."
+                )
+        return values
+
+    def _call(self, inputs: Dict[str, str]) -> Dict[str, str]:
+        _input = inputs[self.input_key]
+        for chain in self.chains:
+            _input = chain.run(_input)
+        return {self.output_key: _input}
