+++ tests/unit_tests/chains/test_pipeline.py
@@ -5,7 +5,7 @@ import pytest
 from pydantic import BaseModel
 
 from langchain.chains.base import Chain
-from langchain.chains.pipeline import Pipeline
+from langchain.chains.sequential import SequentialChain, SimpleSequentialPipeline
 
 
 class FakeChain(Chain, BaseModel):
@@ -32,24 +32,25 @@ class FakeChain(Chain, BaseModel):
         return outputs
 
 
-def test_pipeline_usage_single_inputs() -> None:
-    """Test pipeline on single input chains."""
+def test_sequential_usage_single_inputs() -> None:
+    """Test sequential on single input chains."""
     chain_1 = FakeChain(input_variables=["foo"], output_variables=["bar"])
     chain_2 = FakeChain(input_variables=["bar"], output_variables=["baz"])
-    pipeline = Pipeline(chains=[chain_1, chain_2], input_variables=["foo"])
+    pipeline = SequentialChain(chains=[chain_1, chain_2], input_variables=["foo"])
     output = pipeline({"foo": "123"})
-    expected_output = {"bar": "123foo", "baz": "123foofoo", "foo": "123"}
+    expected_output = {"baz": "123foofoo", "foo": "123"}
     assert output == expected_output
 
 
-def test_pipeline_usage_multiple_inputs() -> None:
-    """Test pipeline on multiple input chains."""
+def test_sequential_usage_multiple_inputs() -> None:
+    """Test sequential on multiple input chains."""
     chain_1 = FakeChain(input_variables=["foo", "test"], output_variables=["bar"])
     chain_2 = FakeChain(input_variables=["bar", "foo"], output_variables=["baz"])
-    pipeline = Pipeline(chains=[chain_1, chain_2], input_variables=["foo", "test"])
+    pipeline = SequentialChain(
+        chains=[chain_1, chain_2], input_variables=["foo", "test"]
+    )
     output = pipeline({"foo": "123", "test": "456"})
     expected_output = {
-        "bar": "123 456foo",
         "baz": "123 456foo 123foo",
         "foo": "123",
         "test": "456",
@@ -57,47 +58,71 @@ def test_pipeline_usage_multiple_inputs() -> None:
     assert output == expected_output
 
 
-def test_pipeline_usage_multiple_outputs() -> None:
-    """Test pipeline usage on multiple output chains."""
+def test_sequential_usage_multiple_outputs() -> None:
+    """Test sequential usage on multiple output chains."""
     chain_1 = FakeChain(input_variables=["foo"], output_variables=["bar", "test"])
     chain_2 = FakeChain(input_variables=["bar", "foo"], output_variables=["baz"])
-    pipeline = Pipeline(chains=[chain_1, chain_2], input_variables=["foo"])
+    pipeline = SequentialChain(chains=[chain_1, chain_2], input_variables=["foo"])
     output = pipeline({"foo": "123"})
     expected_output = {
-        "bar": "123foo",
         "baz": "123foo 123foo",
         "foo": "123",
-        "test": "123foo",
     }
     assert output == expected_output
 
 
-def test_pipeline_missing_inputs() -> None:
+def test_sequential_missing_inputs() -> None:
     """Test error is raised when input variables are missing."""
     chain_1 = FakeChain(input_variables=["foo"], output_variables=["bar"])
     chain_2 = FakeChain(input_variables=["bar", "test"], output_variables=["baz"])
     with pytest.raises(ValueError):
         # Also needs "test" as an input
-        Pipeline(chains=[chain_1, chain_2], input_variables=["foo"])
+        SequentialChain(chains=[chain_1, chain_2], input_variables=["foo"])
 
 
-def test_pipeline_bad_outputs() -> None:
+def test_sequential_bad_outputs() -> None:
     """Test error is raised when bad outputs are specified."""
     chain_1 = FakeChain(input_variables=["foo"], output_variables=["bar"])
     chain_2 = FakeChain(input_variables=["bar"], output_variables=["baz"])
     with pytest.raises(ValueError):
         # "test" is not present as an output variable.
-        Pipeline(
+        SequentialChain(
             chains=[chain_1, chain_2],
             input_variables=["foo"],
             output_variables=["test"],
         )
 
 
-def test_pipeline_overlapping_inputs() -> None:
+def test_sequential_overlapping_inputs() -> None:
     """Test error is raised when input variables are overlapping."""
     chain_1 = FakeChain(input_variables=["foo"], output_variables=["bar", "test"])
     chain_2 = FakeChain(input_variables=["bar"], output_variables=["baz"])
     with pytest.raises(ValueError):
         # "test" is specified as an input, but also is an output of one step
-        Pipeline(chains=[chain_1, chain_2], input_variables=["foo", "test"])
+        SequentialChain(chains=[chain_1, chain_2], input_variables=["foo", "test"])
+
+
+def test_simple_sequential_functionality() -> None:
+    """Test simple sequential functionality."""
+    chain_1 = FakeChain(input_variables=["foo"], output_variables=["bar"])
+    chain_2 = FakeChain(input_variables=["bar"], output_variables=["baz"])
+    pipeline = SimpleSequentialPipeline(chains=[chain_1, chain_2])
+    output = pipeline({"input": "123"})
+    expected_output = {"output": "123foofoo", "input": "123"}
+    assert output == expected_output
+
+
+def test_multi_input_errors() -> None:
+    """Test simple sequential errors if multiple input variables are expected."""
+    chain_1 = FakeChain(input_variables=["foo"], output_variables=["bar"])
+    chain_2 = FakeChain(input_variables=["bar", "foo"], output_variables=["baz"])
+    with pytest.raises(ValueError):
+        SimpleSequentialPipeline(chains=[chain_1, chain_2])
+
+
+def test_multi_output_errors() -> None:
+    """Test simple sequential errors if multiple output variables are expected."""
+    chain_1 = FakeChain(input_variables=["foo"], output_variables=["bar", "grok"])
+    chain_2 = FakeChain(input_variables=["bar"], output_variables=["baz"])
+    with pytest.raises(ValueError):
+        SimpleSequentialPipeline(chains=[chain_1, chain_2])
